backend/requirements.txt
django==3.2.8
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pyotp
cryptography